+++ database/migrations/2018_11_01_175957_create_permission_tables.php
@@ -23,14 +23,14 @@ class CreatePermissionTables extends Migration
         $columnNames = config('permission.column_names');
 
         Schema::create($tableNames['permissions'], function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('guard_name');
             $table->timestamps();
         });
 
         Schema::create($tableNames['roles'], function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('guard_name');
             $table->timestamps();
@@ -40,7 +40,7 @@ class CreatePermissionTables extends Migration
             $tableNames,
             $columnNames
         ) {
-            $table->unsignedInteger('permission_id');
+            $table->unsignedBigInteger('permission_id');
 
             $table->string('model_type', 185);
             $table->unsignedBigInteger($columnNames['model_morph_key']);
@@ -61,7 +61,7 @@ class CreatePermissionTables extends Migration
             $tableNames,
             $columnNames
         ) {
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('role_id');
 
             $table->string('model_type', 185);
             $table->unsignedBigInteger($columnNames['model_morph_key']);
@@ -79,8 +79,8 @@ class CreatePermissionTables extends Migration
         });
 
         Schema::create($tableNames['role_has_permissions'], function (Blueprint $table) use ($tableNames) {
-            $table->unsignedInteger('permission_id');
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('permission_id');
+            $table->unsignedBigInteger('role_id');
 
             $table->foreign('permission_id')
                 ->references('id')
